Impl UserService

+++ src/main/java/entity/Product.java
@@ -1,14 +1,14 @@
 package entity;
 
 public class Product {
-    protected int id;
+    protected long id;
     protected String productName;
     protected float price;
     protected float weight;
     protected String color;
     protected int productCount;
 
-    public Product(int id,String productName,float price, float weight, String color, int productCount){
+    public Product(long id,String productName,float price, float weight, String color, int productCount){
         this.id = id;
         this.productName = productName;
         this.price = price;
@@ -24,7 +24,7 @@ public class Product {
         this.productCount = productCount;
     }
         //Getters
-    public int setId(){
+    public long setId(){
         return id;
     }
