+++ src/main/java/entity/User.java
@@ -1,11 +1,11 @@
 package entity;
 
 public class User {
-    private int id;
+    private long id;
     private String login;
     private String password;
 
-    public User(int id, String login, String password){
+    public User(long id, String login, String password){
         this.id = id ;
         this.login = login;
         this.password = password;
@@ -13,7 +13,7 @@ public class User {
 
     //Getters
 
-    public int getId() {
+    public long getId() {
         return id;
     }
     public String getLogin(){
